# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoInfoInput {
  /** The video URL to fetch info for */
  url: string;
}

/**
 * Whether this is a video or audio format
 */
export type VideoFormatType = typeof VideoFormatType[keyof typeof VideoFormatType];


export const VideoFormatType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface VideoFormat {
  /** Unique format identifier for yt-dlp */
  formatId: string;
  /** Quality identifier e.g. 360p, 720p, 1080p, 4k, mp3_128, mp3_192, mp3_320 */
  quality: string;
  /** Human-readable quality label */
  label: string;
  /** Whether this is a video or audio format */
  type: VideoFormatType;
  /**
     * Approximate file size in bytes, null if unknown
     * @nullable
     */
  filesize: number | null;
  /**
     * Optional badge text e.g. HD, Full HD, 4K, Best Quality
     * @nullable
     */
  badge?: string | null;
}

export interface VideoInfo {
  /** The original URL */
  url: string;
  /** Video title */
  title: string;
  /**
     * Thumbnail URL
     * @nullable
     */
  thumbnail: string | null;
  /**
     * Video duration in seconds, null if unknown
     * @nullable
     */
  duration: number | null;
  /** Detected platform name e.g. YouTube, TikTok */
  platform: string;
  formats: VideoFormat[];
}

export interface DownloadInput {
  /** The video URL */
  url: string;
  /** The format ID to download */
  formatId: string;
}

export interface DownloadResponse {
  /** The URL to download the file from */
  downloadUrl: string;
  /** Suggested filename for the download */
  filename: string;
}

export interface ApiError {
  /** Human-readable error message */
  error: string;
}