# lib/api-zod/src/generated/types/downloadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadInput {
  /** The video URL */
  url: string;
  /** The format ID to download */
  formatId: string;
}