# lib/api-zod/src/generated/types/downloadResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadResponse {
  /** The URL to download the file from */
  downloadUrl: string;
  /** Suggested filename for the download */
  filename: string;
}