# lib/api-zod/src/generated/types/videoFormat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoFormatType } from './videoFormatType';

export interface VideoFormat {
  /** Unique format identifier for yt-dlp */
  formatId: string;
  /** Quality identifier e.g. 360p, 720p, 1080p, 4k, mp3_128, mp3_192, mp3_320 */
  quality: string;
  /** Human-readable quality label */
  label: string;
  /** Whether this is a video or audio format */
  type: VideoFormatType;
  /**
     * Approximate file size in bytes, null if unknown
     * @nullable
     */
  filesize: number | null;
  /**
     * Optional badge text e.g. HD, Full HD, 4K, Best Quality
     * @nullable
     */
  badge?: string | null;
}