# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoFormat } from './videoFormat';

export interface VideoInfo {
  /** The original URL */
  url: string;
  /** Video title */
  title: string;
  /**
     * Thumbnail URL
     * @nullable
     */
  thumbnail: string | null;
  /**
     * Video duration in seconds, null if unknown
     * @nullable
     */
  duration: number | null;
  /** Detected platform name e.g. YouTube, TikTok */
  platform: string;
  formats: VideoFormat[];
}