# lib/api-zod/src/generated/types/videoInfoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VideoTools Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoInfoInput {
  /** The video URL to fetch info for */
  url: string;
}